Fix some navigation problems
fixes #90

In some cases it would seem Svelte renders the page before updating `_page`
so that the machine can't receive the `PAGE_CHANGED` event and
update `_viewData` correctly, so that in some cases the shape of the data
is not what we expect, getting obvious errors.

At least for now the fix is to check that the data we are going to use
refers to the correct `_page` (note that this would fail had we 2 pages needing to
do the same backend query). Hopefully we'll find a better way!

Also:
- removed some unused code
- some refactoring
- renamed `setCxtNextValues` to `setCtxNextValues`